requirements.txt
streamlit==1.22.0
python-dotenv==1.0.0
requests==2.31.0
langchain-google-genai==0.1.0  
google-generativeai==0.4.0     
pydantic==2.5.3      
langgraph          